migrate dashboard data layer: Neon Postgres -> Cloudflare D1 + R2
- functions/_lib/db.ts: D1-backed tagged-template shim (was neon HTTP)
- functions/_lib/auth.ts Env: DB(D1Database)+VOICE_BUCKET(R2)+VOICE_PUBLIC_URL
- all 22 API handlers: SQLite dialect (drop ::casts, now()->strftime, json_extract, JSON.parse on read)
- voice ingest->R2 put, clip serves from R2 (was Postgres bytea)
- sql/d1/schema.sql (15 tables) + schema_reconstructed.sql (4 handler-referenced tables)
- wrangler.toml: D1+R2 bindings, VOICE_PUBLIC_URL; DATABASE_URL removed
- scripts: _d1.ts REST shim, sync-from-pai -> D1, seed-099.ts, reingest-voice.ts

+++ functions/api/agents/index.ts
@@ -51,7 +51,7 @@ export const onRequestGet: PagesFunction<Env> = async ({ request, env }) => {
         FROM tickets t
         WHERE t.agent = a.name
           AND t.status IN ('todo','in_progress','review')
-      )::int AS in_flight_count,
+      ) AS in_flight_count,
       a.updated_at
     FROM agents a
     LEFT JOIN tickets ct ON ct.id = a.current_ticket_id
